refactor: centralize conversation uuid resolution
