+++ apps/shared/lib/uuid.ts
@@ -1,3 +1,5 @@
+import { isUuid as jsIsUuid } from '../../../packages/conversation-uuid/index.js';
+
 export function isUuid(v: string): boolean {
-  return /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i.test(v ?? '');
+  return jsIsUuid(v);
 }
