+++ lib/alertConversation.js
@@ -1,6 +1,8 @@
 import { buildAlertConversationLink, normalizeAlertLinkInput } from './conversationLink.js';
-import { resolveConversationUuidHedged as defaultResolveConversationUuid } from '../apps/shared/lib/conversationUuid.js';
-import { mintUuidFromRaw } from '../apps/shared/lib/canonicalConversationUuid.js';
+import {
+  resolveConversationUuidHedged as defaultResolveConversationUuid,
+  mintUuidFromRaw,
+} from '../packages/conversation-uuid/index.js';
 
 const UUID_RE = /[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}/i;
 const LEGACY_RE = /^\d+$/;
