+++ lib/alertLink.js
@@ -1,6 +1,6 @@
 import { appUrl, makeConversationLink } from './links.js';
 import { makeLinkToken } from './linkToken.js';
-import { mintUuidFromRaw } from '../apps/shared/lib/canonicalConversationUuid.js';
+import { mintUuidFromRaw } from '../packages/conversation-uuid/index.js';
 import { resolveViaInternalEndpointWithDetails } from './internalResolve.js';
 
 let cachedTryResolveConversationUuid =
@@ -24,7 +24,7 @@ async function getTryResolveConversationUuid() {
   attemptedDynamicTryResolveLoad = true;
 
   try {
-    const mod = await import('../apps/server/lib/conversations.js');
+    const mod = await import('../packages/conversation-uuid/index.js');
     const fn = mod?.tryResolveConversationUuid;
     if (typeof fn === 'function') {
       cachedTryResolveConversationUuid = fn;
@@ -51,7 +51,7 @@ async function getResolveConversationUuid() {
   if (attemptedDynamicResolveLoad) return null;
   attemptedDynamicResolveLoad = true;
   try {
-    const mod = await import('../apps/shared/lib/conversationUuid.js');
+    const mod = await import('../packages/conversation-uuid/index.js');
     const fn = mod?.resolveConversationUuid;
     if (typeof fn === 'function') {
       cachedResolveConversationUuid = fn;
