+++ packages/conversation-uuid/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "@boom/conversation-uuid",
+  "type": "module"
+}
