+++ packages/conversation-uuid/src/utils.js
@@ -1,0 +1,18 @@
+const UUID_RE = /^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+
+export function isUuid(value) {
+  return UUID_RE.test(String(value ?? ''));
+}
+
+export function normalizeIdentifier(raw) {
+  if (raw == null) return '';
+  if (typeof raw === 'string') return raw.trim();
+  if (typeof raw === 'number' && Number.isFinite(raw)) return String(Math.trunc(raw));
+  if (typeof raw === 'bigint') return raw.toString();
+  return String(raw).trim();
+}
+
+export const __test__ = {
+  UUID_RE,
+  normalizeIdentifier,
+};
